+++ src/main/java/cart/global/exception/cart/ProductNotFoundInCartException.java
@@ -1,7 +1,7 @@
 package cart.global.exception.cart;
 
-import cart.global.exception.CartException;
-import cart.global.exception.ExceptionStatus;
+import cart.global.exception.common.CartException;
+import cart.global.exception.common.ExceptionStatus;
 
 public class ProductNotFoundInCartException extends CartException {
     public ProductNotFoundInCartException() {
