+++ src/main/java/cart/global/exception/common/ExceptionStatus.java
@@ -1,4 +1,4 @@
-package cart.global.exception;
+package cart.global.exception.common;
 
 import org.springframework.http.HttpStatus;
